+++ crates/sdk-core/tests/integ_tests/heartbeat_tests.rs
@@ -13,13 +13,12 @@ use temporalio_common::protos::{
         workflow_commands::{ActivityCancellationType, ScheduleActivity},
         workflow_completion::WorkflowActivationCompletion,
     },
-    temporal::api::{
-        common::v1::{Payload, RetryPolicy},
-        enums::v1::TimeoutType,
-    },
+    temporal::api::common::v1::{Payload, RetryPolicy},
 };
 use temporalio_macros::{workflow, workflow_methods};
-use temporalio_sdk::{ActivityExecutionError, ActivityOptions, WorkflowContext, WorkflowResult};
+use temporalio_sdk::{
+    ActivityExecutionError, ActivityOptions, TimeoutType, WorkflowContext, WorkflowResult,
+};
 use temporalio_sdk_core::{
     prost_dur,
     replay::DEFAULT_ACTIVITY_TYPE,